requirements.txt
selenium
webdriver-manager
requests
beautifulsoup4
groq
python-slugify
jinja2
